requirements.txt
flask
requests
beautifulsoup4
python-dotenv
google-cloud-texttospeech
google-generativeai
tqdm
urllib3
